commands: Moved files code into 'commands/files' subpackage
